Splitter opp EuxService
Alt som handler om prefill/opprettelse forblir i EuxService
Alt som handler om innhenting av informasjon flyttes til EuxInnhentingService.kt

Ubrukt kode slettes
