+++ src/main/kotlin/no/nav/eessi/pensjon/fagmodul/prefill/InnhentingService.kt
@@ -1,16 +1,12 @@
 package no.nav.eessi.pensjon.fagmodul.prefill
 
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry
-import no.nav.eessi.pensjon.fagmodul.eux.BucUtils
-import no.nav.eessi.pensjon.fagmodul.eux.EuxService
 import no.nav.eessi.pensjon.fagmodul.models.PersonDataCollection
 import no.nav.eessi.pensjon.fagmodul.models.PrefillDataModel
 import no.nav.eessi.pensjon.metrics.MetricsHelper
 import no.nav.eessi.pensjon.personoppslag.pdl.model.IdentType
 import no.nav.eessi.pensjon.personoppslag.pdl.model.NorskIdent
 import no.nav.eessi.pensjon.vedlegg.VedleggService
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
@@ -18,11 +14,8 @@ import javax.annotation.PostConstruct
 
 @Service
 class InnhentingService(private val personDataService: PersonDataService,
-                        private val euxService: EuxService,
                         private val vedleggService: VedleggService,
                         @Autowired(required = false) private val metricsHelper: MetricsHelper = MetricsHelper(SimpleMeterRegistry())) {
-
-    private val logger: Logger = LoggerFactory.getLogger(InnhentingService::class.java)
 
     private lateinit var HentPerson: MetricsHelper.Metric
     private lateinit var addInstutionAndDocumentBucUtils: MetricsHelper.Metric
@@ -35,18 +28,6 @@ class InnhentingService(private val personDataService: PersonDataService,
             "AddInstutionAndDocumentBucUtils",
             ignoreHttpCodes = listOf(HttpStatus.BAD_REQUEST)
         )
-    }
-
-    fun kanSedOpprettes(dataModel: PrefillDataModel): BucUtils {
-        return addInstutionAndDocumentBucUtils.measure {
-            logger.info("******* Hent BUC sjekk om sed kan opprettes *******")
-            BucUtils(euxService.getBuc(dataModel.euxCaseID)).also { bucUtil ->
-                //sjekk for om deltakere alt er fjernet med x007 eller x100 sed
-                bucUtil.checkForParticipantsNoLongerActiveFromXSEDAsInstitusjonItem(dataModel.getInstitutionsList())
-                //gyldig sed kan opprettes
-                bucUtil.checkIfSedCanBeCreated(dataModel.sedType, dataModel.penSaksnummer)
-            }
-        }
     }
 
     fun hentPersonData(prefillData: PrefillDataModel) : PersonDataCollection = personDataService.hentPersonData(prefillData)
